Change UnifyVcf script to be called through different wrappers reflecting the different suppliers (Dragen, ONT, PacBio)
